Move the parser packages around
